scala 3.1.2

+++ js/src/main/scala/no/skytteren/scalatime/scalatime.scala
@@ -20,7 +20,7 @@ package object scalatime {
     def now(): DateTime = {
       val jDateTime = new scalajs.js.Date()
       new DateTime(
-        Year(jDateTime.getFullYear()), Month(jDateTime.getMonth().toShort + 1), DayOfMonth(jDateTime.getDate().toShort),
+        Year(jDateTime.getFullYear().toLong), Month(jDateTime.getMonth().toShort + 1), DayOfMonth(jDateTime.getDate().toShort),
         Hour(jDateTime.getHours.toShort), Minute(jDateTime.getMinutes.toShort), Second(jDateTime.getSeconds.toShort), Millisecond(jDateTime.getMilliseconds().toShort)
       )
     }
